Split monster RepositoryData controller into 4 separate controllers
This looks rather more organised, although it does involve stashing
some shared functions in a trait, which I don't love.
