+++ modules/admin/app/views/admin/repository/adminActions.scala.html
@@ -45,7 +45,7 @@
     @views.html.common.sidebarSection(Messages("ingest")) {
         @views.html.admin.common.sidebarActionGroup {
             @views.html.admin.common.sidebarAction(user.hasPermission(ContentTypes.Repository, PermissionType.Update)) {
-                <a href="@controllers.institutions.routes.RepositoryData.manager(item.id)">@Messages("repository.data.manage")</a>
+                <a href="@controllers.institutions.routes.ImportFiles.manager(item.id)">@Messages("repository.data.manage")</a>
             }
             @views.html.admin.common.sidebarAction(userOpt.forall(_.isAdmin)) {
                 <a href="@controllers.institutions.routes.Repositories.ingest(item.id)">@Messages("ingest.format.ead")</a>
