+++ modules/admin/app/views/admin/repository/datamanager.scala.html
@@ -26,42 +26,42 @@
         <script src="@controllers.portal.routes.PortalData.jsRoutes()"></script>
         @helper.javascriptRouter("dataManagerRoutes")(
             controllers.admin.routes.javascript.Tasks.taskMonitorWS,
-            controllers.institutions.routes.javascript.RepositoryData.listFiles,
-            controllers.institutions.routes.javascript.RepositoryData.info,
-            controllers.institutions.routes.javascript.RepositoryData.validateFiles,
-            controllers.institutions.routes.javascript.RepositoryData.validateAll,
-            controllers.institutions.routes.javascript.RepositoryData.deleteFiles,
-            controllers.institutions.routes.javascript.RepositoryData.deleteAll,
-            controllers.institutions.routes.javascript.RepositoryData.ingestFiles,
-            controllers.institutions.routes.javascript.RepositoryData.ingestAll,
-            controllers.institutions.routes.javascript.RepositoryData.uploadHandle,
-            controllers.institutions.routes.javascript.RepositoryData.fileUrls,
-            controllers.institutions.routes.javascript.RepositoryData.harvestOaiPmh,
-            controllers.institutions.routes.javascript.RepositoryData.cancelOaiPmhHarvest,
-            controllers.institutions.routes.javascript.RepositoryData.getOaiPmhConfig,
-            controllers.institutions.routes.javascript.RepositoryData.saveOaiPmhConfig,
-            controllers.institutions.routes.javascript.RepositoryData.deleteOaiPmhConfig,
-            controllers.institutions.routes.javascript.RepositoryData.testOaiPmhConfig,
-            controllers.institutions.routes.javascript.RepositoryData.convertFile,
-            controllers.institutions.routes.javascript.RepositoryData.convert,
-            controllers.institutions.routes.javascript.RepositoryData.cancelConvert,
-            controllers.institutions.routes.javascript.RepositoryData.getConvertConfig,
-            controllers.institutions.routes.javascript.RepositoryData.saveConvertConfig,
-            controllers.institutions.routes.javascript.RepositoryData.listDataTransformations,
-            controllers.institutions.routes.javascript.RepositoryData.getDataTransformation,
-            controllers.institutions.routes.javascript.RepositoryData.createDataTransformation,
-            controllers.institutions.routes.javascript.RepositoryData.updateDataTransformation,
-            controllers.institutions.routes.javascript.RepositoryData.deleteDataTransformation,
-            controllers.institutions.routes.javascript.RepositoryData.listDatasets,
-            controllers.institutions.routes.javascript.RepositoryData.datasetStats,
-            controllers.institutions.routes.javascript.RepositoryData.createDataset,
-            controllers.institutions.routes.javascript.RepositoryData.updateDataset,
-            controllers.institutions.routes.javascript.RepositoryData.deleteDataset
+            controllers.institutions.routes.javascript.ImportFiles.listFiles,
+            controllers.institutions.routes.javascript.ImportFiles.info,
+            controllers.institutions.routes.javascript.ImportFiles.validateFiles,
+            controllers.institutions.routes.javascript.ImportFiles.validateAll,
+            controllers.institutions.routes.javascript.ImportFiles.deleteFiles,
+            controllers.institutions.routes.javascript.ImportFiles.deleteAll,
+            controllers.institutions.routes.javascript.ImportFiles.ingestFiles,
+            controllers.institutions.routes.javascript.ImportFiles.ingestAll,
+            controllers.institutions.routes.javascript.ImportFiles.uploadHandle,
+            controllers.institutions.routes.javascript.ImportFiles.fileUrls,
+            controllers.institutions.routes.javascript.OaiPmhConfigs.harvest,
+            controllers.institutions.routes.javascript.OaiPmhConfigs.cancelHarvest,
+            controllers.institutions.routes.javascript.OaiPmhConfigs.get,
+            controllers.institutions.routes.javascript.OaiPmhConfigs.save,
+            controllers.institutions.routes.javascript.OaiPmhConfigs.delete,
+            controllers.institutions.routes.javascript.OaiPmhConfigs.test,
+            controllers.institutions.routes.javascript.DataTransformations.convertFile,
+            controllers.institutions.routes.javascript.DataTransformations.convert,
+            controllers.institutions.routes.javascript.DataTransformations.cancelConvert,
+            controllers.institutions.routes.javascript.DataTransformations.getConfig,
+            controllers.institutions.routes.javascript.DataTransformations.saveConfig,
+            controllers.institutions.routes.javascript.DataTransformations.list,
+            controllers.institutions.routes.javascript.DataTransformations.get,
+            controllers.institutions.routes.javascript.DataTransformations.create,
+            controllers.institutions.routes.javascript.DataTransformations.update,
+            controllers.institutions.routes.javascript.DataTransformations.delete,
+            controllers.institutions.routes.javascript.ImportDatasets.list,
+            controllers.institutions.routes.javascript.ImportDatasets.stats,
+            controllers.institutions.routes.javascript.ImportDatasets.create,
+            controllers.institutions.routes.javascript.ImportDatasets.update,
+            controllers.institutions.routes.javascript.ImportDatasets.delete
 
         )
 
         <script>
-                window.SERVICE = dataManagerRoutes.controllers.institutions.RepositoryData;
+                window.SERVICE = dataManagerRoutes.controllers.institutions;
                 window.CONFIG = {
                     repositoryId: "@item.id",
                     versioned: @isVersioned,
